Implement the code generator in template haskell
As well as:

- Add a function to generate games using quoted terms rather than strings.
- Remove sources from examples and inline them in their files using
  the new Template Haskell code generator.
- Add quickcheck tests to ensure the functionality of the new code
  generator is the same as the previous one.

The code generator now takes a `QBlock` and a list of `QLine` which
allow to generate the functions for games without having to copy
and paste a string in to the source code. This also makes
checked-in code much cleaner and easier to read.

Code is generated with `generateGame` and takes either a `QBlock` or
a list of `QLine`. `generateGame` has the following signature:

generateGame :: GameCompiler ast -> String -> [String] -> ast -> Q [Dec]

The first argument allows us to pick the name of the function we
want Template Haskell to generate for us, the second one is the
list of arguments that the function uses. Game-theoretically, those
are exogenous parameters. `ast` is constrained on `GameCompiler` which
ensures there is a way to compile the given AST into a Template Haskell
definition.